Renamed to fit spec requirements, changed figure export path
